% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{compute_ecm}
\alias{compute_ecm}
\title{Compute ECM Parameterization}
\usage{
compute_ecm(est, y, X, p, q, tau, constant = TRUE)
}
\arguments{
\item{est}{QARDL estimation results.}

\item{y}{Dependent variable vector.}

\item{X}{Covariate matrix.}

\item{p}{AR lag order.}

\item{q}{Distributed lag order.}

\item{tau}{Vector of quantiles.}

\item{constant}{Logical for intercept.}
}
\value{
List containing ECM parameters.
}
\description{
Computes Error Correction Model parameters.
}
\keyword{internal}
