% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{compute_longrun}
\alias{compute_longrun}
\title{Compute Long-Run Parameters}
\usage{
compute_longrun(est, k, tau, y, X, constant = TRUE)
}
\arguments{
\item{est}{QARDL estimation results from \code{qardl_estimate}.}

\item{k}{Number of covariates.}

\item{tau}{Vector of quantiles.}
}
\value{
List containing beta, beta_se, beta_cov, rho, rho_se.
}
\description{
Computes long-run (beta) parameters from QARDL estimates.
}
\keyword{internal}
