% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_simulate.R
\name{dgp_qardl}
\alias{dgp_qardl}
\title{Data Generating Process for QARDL}
\usage{
dgp_qardl(nobs, p, q, k, phi, gamma, sigma_u, sigma_x)
}
\arguments{
\item{nobs}{Sample size.}

\item{p}{AR lag order.}

\item{q}{Distributed lag order.}

\item{k}{Number of covariates.}

\item{phi}{AR parameters.}

\item{gamma}{Impact parameters.}

\item{sigma_u}{Error standard deviation.}

\item{sigma_x}{Covariate innovation standard deviation.}
}
\value{
List with y and X.
}
\description{
Internal function to generate data from a QARDL DGP.
}
\keyword{internal}
