% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_print.R
\name{print.qardl}
\alias{print.qardl}
\title{Print QARDL Results}
\usage{
\method{print}{qardl}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"qardl"}.}

\item{digits}{Number of decimal places. Default is 4.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisible \code{x}.
}
\description{
Print method for QARDL estimation results.
}
