% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_simulate.R
\name{print.qardl_mc}
\alias{print.qardl_mc}
\title{Print Monte Carlo Results}
\usage{
\method{print}{qardl_mc}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{"qardl_mc"}.}

\item{digits}{Number of decimal places. Default is 4.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisible \code{x}.
}
\description{
Print Monte Carlo Results
}
