% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_wald.R
\name{print.qardl_wald}
\alias{print.qardl_wald}
\title{Print QARDL Wald Test Results}
\usage{
\method{print}{qardl_wald}(x, digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{"qardl_wald"}.}

\item{digits}{Number of decimal places. Default is 4.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisible \code{x}.
}
\description{
Print QARDL Wald Test Results
}
