% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_bic.R
\name{print_bic_grid}
\alias{print_bic_grid}
\title{Print BIC Grid}
\usage{
print_bic_grid(bic_result, digits = 3)
}
\arguments{
\item{bic_result}{Result from \code{qardl_bic_select}.}

\item{digits}{Number of decimal places. Default is 3.}
}
\value{
Invisible \code{NULL}. Called for side effect of printing.
}
\description{
Prints a formatted BIC grid for lag selection.
}
\examples{
data(qardl_sim)
y <- qardl_sim$y
X <- as.matrix(qardl_sim[, c("x1", "x2")])
bic_result <- qardl_bic_select(y, X, pmax = 4, qmax = 4)
print_bic_grid(bic_result)

}
