% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl.R
\name{qardl_estimate}
\alias{qardl_estimate}
\title{Core QARDL Estimation}
\usage{
qardl_estimate(y, X, p, q, tau, constant = TRUE)
}
\arguments{
\item{y}{Numeric vector of dependent variable.}

\item{X}{Matrix of covariates.}

\item{p}{AR lag order.}

\item{q}{Distributed lag order.}

\item{tau}{Vector of quantiles.}

\item{constant}{Logical for intercept inclusion.}
}
\value{
List of estimated parameters and their covariances.
}
\description{
Internal function for QARDL parameter estimation.
}
\keyword{internal}
