% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_wald.R
\name{wald_pairwise_tests}
\alias{wald_pairwise_tests}
\title{Pairwise Wald Tests}
\usage{
wald_pairwise_tests(object, type)
}
\arguments{
\item{object}{QARDL object.}

\item{type}{Parameter type.}
}
\value{
Data frame of pairwise test results.
}
\description{
Internal function for pairwise parameter equality tests.
}
\keyword{internal}
