% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_bic.R
\name{qardl_bic_select}
\alias{qardl_bic_select}
\title{BIC-Based Lag Order Selection for QARDL}
\usage{
qardl_bic_select(y, X, pmax = 7L, qmax = 7L, constant = TRUE)
}
\arguments{
\item{y}{Numeric vector of dependent variable.}

\item{X}{Matrix of covariates.}

\item{pmax}{Integer. Maximum AR lag order to consider. Default is 7.}

\item{qmax}{Integer. Maximum distributed lag order to consider. Default is 7.}

\item{constant}{Logical. Include intercept. Default is \code{TRUE}.}
}
\value{
A list containing:
\describe{
  \item{p_opt}{Optimal AR lag order}
  \item{q_opt}{Optimal distributed lag order}
  \item{bic_grid}{Matrix of BIC values (pmax x qmax)}
  \item{bic_min}{Minimum BIC value}
}
}
\description{
Automatically selects optimal lag orders (p, q) for the QARDL model
using the Bayesian Information Criterion (BIC) evaluated at the median
quantile (tau = 0.5).
}
\details{
The BIC is computed using the Schwarz criterion at the median quantile:
\deqn{BIC(p, q) = \log(\hat{\sigma}^2_{\tau=0.5}) + \frac{k_{pq} \log(n)}{n}}

where \eqn{k_{pq}} is the number of parameters (p AR terms + q*k impact terms
+ constant) and \eqn{\hat{\sigma}^2} is the estimated residual variance.
}
\examples{
data(qardl_sim)
y <- qardl_sim$y
X <- as.matrix(qardl_sim[, c("x1", "x2")])
bic_result <- qardl_bic_select(y, X, pmax = 5, qmax = 5)
print(bic_result$bic_grid)

}
\references{
Cho, J.S., Kim, T.-H., & Shin, Y. (2015). Quantile cointegration in the
autoregressive distributed-lag modeling framework. \emph{Journal of
Econometrics}, 188(1), 281-300. \doi{10.1016/j.jeconom.2015.01.003}
}
\seealso{
\code{\link{qardl}}
}
