% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qardl_wald.R
\name{wald_constancy_test}
\alias{wald_constancy_test}
\title{Wald Constancy Test}
\usage{
wald_constancy_test(params, cov_array, nobs, param_name)
}
\arguments{
\item{params}{Parameter matrix (dim x ntau).}

\item{cov_array}{Covariance array (dim x dim x ntau).}

\item{nobs}{Number of observations.}

\item{param_name}{Name of parameter for labeling.}
}
\value{
List with statistic, df, pvalue.
}
\description{
Internal function to perform Wald test for parameter constancy.
}
\keyword{internal}
